requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
drf-spectacular>=0.27
psycopg[binary]>=3.1
scikit-learn>=1.5
pandas>=2.2
joblib>=1.4
python-dotenv>=1.0
dj-database-url>=2.2
argon2-cffi>=23.1.0